refactored esbuild building
