+++ .gitignore
@@ -4,4 +4,4 @@ _site
 .jekyll-cache
 .jekyll-metadata
 vendor
-/site/assets/dist
+/site/assets/js/
