+++ esbuild.config.js
@@ -1,10 +1,10 @@
 const esbuild = require('esbuild');
 
 esbuild.build({
-  entryPoints: ['./assets/js/index.js'],
+  entryPoints: ['./src/index.js'],
   bundle: true,
   minify: true,
   sourcemap: true,
   target: ['es2018'],
-  outfile: './assets/dist/bundle.js',
+  outfile: './site/assets/js/script.js',
 }).catch(() => process.exit(1));
