+++ esbuild.watch.cjs
@@ -1,13 +1,12 @@
 const esbuild = require('esbuild');
 
-// Start esbuild with watch mode
 const ctx = esbuild.context({
   entryPoints: ['./src/index.js'],
   bundle: true,
-  minify: false,
+  minify: true,
   sourcemap: true,
   target: ['es2018'],
-  outfile: './site/assets/dist/bundle.js',
+  outfile: './site/assets/js/script.js',
 }).catch(() => process.exit(1));
 
 ctx.then(context => {
