+++ site/_layouts/default.html
@@ -57,8 +57,7 @@
       href="{{ site.url }}/assets/favicons/apple-touch-icon.png"
     />
     <script src="https://www.google.com/recaptcha/api.js" async defer></script>
-    <script src="{{ site.url }}/assets/js/recaptcha.js"></script>
-    <script src="{{ site.url }}/assets/dist/bundle.js"></script>
+    <script src="{{ site.url }}/assets/js/script.js"></script>
   </head>
   <body class="body-gradient body-marketing">
     <div class="page">
