+++ src/currency-setter.js
