+++ src/index.js
@@ -1,12 +1,8 @@
 import '@tabler/core/'
 
-import '../site/assets/js/theme-utils.js'
-import '../site/assets/js/theme-initialization.js'
-import '../site/assets/js/theme-toggler.js'
-import '../site/assets/js/currency-setter.js'
-import '../site/assets/js/intl-format-number-handler.js'
-import '../site/assets/js/recaptcha.js'
-
-document.addEventListener('DOMContentLoaded', () => {
-  console.log('Application initialized with esbuild');
-});
+import './theme-utils.js'
+import './theme-initialization.js'
+import './theme-toggler.js'
+import './currency-setter.js'
+import './intl-format-number-handler.js'
+import './recaptcha.js'
