+++ src/intl-format-number-handler.js
